pyproject.toml
[project]
name = "azure-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "azure-search-documents>=11.6.0",
    "chainlit>=2.9.6",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.131.0",
    "greenlet>=3.3.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langsmith>=0.7.6",
    "markdown>=3.10.2",
    "mcp>=1.26.0",
    "nemoguardrails>=0.20.0",
    "presidio-analyzer>=2.2.361",
    "presidio-anonymizer>=2.2.361",
    "psycopg-binary>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.2",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.22",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "redis[hiredis]>=7.3.0",
    "redisvl>=0.3.9",
    "spacy>=3.8.11",
    "unstructured>=0.21.5",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
